requirements.txt
fastapi
uvicorn[standard]
tensorflow-cpu
python-multipart
pillow
jinja2
